Moved the statistics tests in stats_test
